api/requirements.txt

fastapi
pydantic
sqlmodel
uvicorn
numpy
tensorflow
pillow